refactor code for readablity
